requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.9.0
